Cleaned up utils.h
