+++ .vscode/tasks.json
@@ -11,8 +11,10 @@
                 "${workspaceFolder}\\src\\creature\\*.cpp",
                 "${workspaceFolder}\\src\\creature\\main_character\\*.cpp",
                 "${workspaceFolder}\\src\\start_menu\\*.cpp",
-                "${workspaceFolder}\\src\\utils\\*.cpp",
+                "${workspaceFolder}\\src\\utils\\customize_text\\*.cpp",
                 "${workspaceFolder}\\src\\utils\\get_user_input\\*.cpp",
+                "${workspaceFolder}\\src\\utils\\handle_sound\\*.cpp",
+                "${workspaceFolder}\\src\\utils\\print\\*.cpp",
                 "${workspaceFolder}\\src\\utils\\roll_dice\\*.cpp",
                 "-lWinmm",
                 "-o",
