+++ src/start_menu/start_menu.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "../utils/utils.h"
 
 namespace start_menu
 {
