+++ src/utils/handle_sound/handle_sound.cpp
@@ -1,40 +1,6 @@
-#include "utils.h"
+#include "handle_sound.h"
 #include <iostream>
-#include <string>
-#include <thread>
-#include <chrono>
-#include <cwchar>
 #include <windows.h>
-
-int utils::g_sleep_for_ms = 100;
-int utils::g_text_size = 16;
-utils::FontWeightValues utils::g_text_weight = utils::FontWeightValues::Normal;
-
-void utils::Print(std::initializer_list<std::string> list, int sleep_for_ms)
-{
-	for (auto string : list)
-	{
-		for (auto character : string)
-		{
-			std::cout << character;
-			std::this_thread::sleep_for(std::chrono::milliseconds(sleep_for_ms));
-		}
-	}
-	std::cout << std::endl;
-}
-
-void utils::CustomizeText(int font_size, utils::FontWeightValues font_weight, const wchar_t* face_name)
-{
-	CONSOLE_FONT_INFOEX cfi;
-	cfi.cbSize = sizeof(cfi);
-	cfi.nFont = 0; // The index of the font in the system's console font table.
-	cfi.dwFontSize.X = 0; // A COORD structure that contains the width and height of each character in the font.
-	cfi.dwFontSize.Y = font_size; //   The X member contains the width, while the Y member contains the height.
-	cfi.FontFamily = FF_DONTCARE; // The font pitch and family. See also TEXTMETRICA.
-	cfi.FontWeight = font_weight; // The font weight. The weight can range from 100 to 1000, in multiples of 100.
-	std::wcscpy(cfi.FaceName, face_name); // Copies the desired typeface over to FaceName.
-	SetCurrentConsoleFontEx(GetStdHandle(STD_OUTPUT_HANDLE), FALSE, &cfi);
-}
 
 void utils::HandleSound(utils::SoundOperations operation)
 {
